Move _get_obj_attrs_map into data_info
